Introduces a configuration loader
The configuration loader loads yaml configuration files. An optional
execution environment can be specified.

In a config file, the 'environment' property specifies different
environments that can be swapped when the file is loaded based
on the environment specified in the config loader.

This replacement is simple for now, but can become more sophisticated
as the need arises.

Change-Id: I4bfdcc0d8dadb7844b715c297a018edc2dc9b19f